Reordering C common utils and adding daemon utils
