+++ src/agents/pnp/PnpAgent.c
@@ -337,53 +337,17 @@ static void ForkDaemon()
     }
 }
 
-bool IsPlatformActive(void)
-{
-   const char* isPlatformActive = "systemctl is-active osconfig-platform";
-   bool status = true;
-
-   if (ESRCH == ExecuteCommand(NULL, isPlatformActive, false, false, 0, 0, NULL, NULL, GetLog())) 
-   {
-        status = false;
-   }
-   
-   return status;
-}
-
-static bool StartPlatform(void)
-{
-   const char* enablePlatform = "systemctl enable osconfig-platform";
-   const char* startPlatform = "systemctl start osconfig-platform";
-   bool status = true;
-
-   if (false == IsPlatformActive()) 
-   {
-        OsConfigLogInfo(GetLog(), "Start the platform");
-        status = ((0 == ExecuteCommand(NULL, enablePlatform, false, false, 0, 0, NULL, NULL, GetLog())) &&
-            (0 == ExecuteCommand(NULL, startPlatform, false, false, 0, 0, NULL, NULL, GetLog())));
-   }
-   
-   return status;
-}
-
-static bool StopPlatform(void)
-{
-   const char* stopPlatform = "sudo systemctl stop osconfig-platform";
-
-   return (0 == ExecuteCommand(NULL, stopPlatform, false, false, 0, 0, NULL, NULL, GetLog()));
-}
-
 bool RefreshMpiClientSession(void)
 {
     bool status = true;
 
-    if (IsPlatformActive())
+    if (IsDaemonActive(OSCONFIG_PLATFORM, GetLog()))
     {
         // Platform is already running
         return status;
     }
     
-    if (true == (status = StartPlatform()))
+    if (true == (status = EnableAndStartDaemon(OSCONFIG_PLATFORM, GetLog())))
     {
         sleep(1);
         
@@ -455,22 +419,10 @@ static void SaveReportedConfigurationToFile()
     int payloadSizeBytes = 0;
     size_t payloadHash = 0;
     int mpiResult = MPI_OK;
-    if (g_localManagement)
+    if (g_localManagement && RefreshMpiClientSession())
     {
         mpiResult = CallMpiGetReported((MPI_JSON_STRING*)&payload, &payloadSizeBytes);
-
-        if ((MPI_OK != mpiResult) && (false == IsPlatformActive()))
-        {
-            CallMpiFree(payload);
-
-            // Restart the platform, re-open the MPI session and retry MpiGetReported
-            if (RefreshMpiClientSession())
-            {
-                mpiResult = CallMpiGetReported((MPI_JSON_STRING*)&payload, &payloadSizeBytes);
-            }
-        }
-
-
+        
         if ((MPI_OK == mpiResult) && (NULL != payload) && (0 < payloadSizeBytes))
         {
             if (g_reportedHash != (payloadHash = HashString(payload)))
@@ -513,7 +465,7 @@ static void LoadDesiredConfigurationFromFile()
     RestrictFileAccessToCurrentAccountOnly(DC_FILE);
 
     payload = LoadStringFromFile(DC_FILE, false, GetLog());
-    if (payload && (0 != (payloadSizeBytes = strlen(payload))))
+    if (payload && (0 != (payloadSizeBytes = strlen(payload)) && RefreshMpiClientSession()))
     {
         // Do not call MpiSetDesired unless this desired configuration is different from previous
         if (g_desiredHash != (payloadHash = HashString(payload)))
@@ -819,6 +771,8 @@ done:
     FREE_MEMORY(connectionString);
     FREE_MEMORY(g_iotHubConnectionString);
 
+    StopAndDisableDaemon(OSCONFIG_PLATFORM, GetLog());
+
     CloseAgent();
     CloseTraceLogging();
     CloseLog(&g_agentLog);
@@ -839,9 +793,6 @@ done:
     {
         free((void *)g_proxyOptions.password);
     }
-
-    // Stop the platform when agent terminates
-    StopPlatform();
 
     return 0;
 }
