+++ src/agents/pnp/PnpUtils.c
@@ -539,19 +539,14 @@ IOTHUB_CLIENT_RESULT ReportPropertyToIotHub(const char* componentName, const cha
         return IOTHUB_CLIENT_ERROR;
     }
 
-    mpiResult = CallMpiGet(componentName, propertyName, &valuePayload, &valueLength);
-
-    if ((MPI_OK != mpiResult) && (false == IsPlatformActive()))
+    if (false == RefreshMpiClientSession())
     {
-        CallMpiFree(valuePayload);
-
-        // Restart the platform, re-open the MPI session and retry MpiGet
-        if (RefreshMpiClientSession())
-        {
-            mpiResult = CallMpiGet(componentName, propertyName, &valuePayload, &valueLength);
-        }
+        LogErrorWithTelemetry(GetLog(), "%s: cannot reach the platform to get reported data for the property %s", componentName, propertyName);
+        return IOTHUB_CLIENT_ERROR;
     }
 
+    mpiResult =  CallMpiGet(componentName, propertyName, &valuePayload, &valueLength);
+        
     if ((MPI_OK == mpiResult) && (valueLength > 0) && (NULL != valuePayload))
     {
         decoratedLength = strlen(componentName) + strlen(propertyName) + valueLength + EXTRA_PROP_PAYLOAD_ESTIMATE;
@@ -648,6 +643,11 @@ IOTHUB_CLIENT_RESULT UpdatePropertyFromIotHub(const char* componentName, const c
     if (NULL == serializedValue)
     {
         OsConfigLogInfo(GetLog(), "%s: %s property update requested with no data (nothing to do)", componentName, propertyName);
+    }
+    else if (false == RefreshMpiClientSession())
+    {
+        LogErrorWithTelemetry(GetLog(), "%s: cannot reach the platform to apply desired data for property %s", componentName, propertyName);
+        mpiResult = ECONNREFUSED;
     }
     else
     {
