+++ src/agents/pnp/inc/AgentCommon.h
@@ -47,6 +47,8 @@
 #include <Mpi.h>
 #include <version.h>
 
+#define OSCONFIG_PLATFORM "osconfig-platform"
+
 // Max number of bytes allowed to go through to Twins (4KB)
 #define OSCONFIG_MAX_PAYLOAD 4096
