+++ src/common/commonutils/CMakeLists.txt
@@ -4,11 +4,23 @@
 project(commonutils)
 
 find_package(RapidJSON REQUIRED)
-add_library(commonutils STATIC CommonUtils.c CommonUtils.cpp)
+
+add_library(commonutils STATIC 
+    CommandUtils.c
+    DeviceInfoUtils.c
+    FileUtils.c
+    OtherUtils.c
+    ProxyUtils.c
+    SocketUtils.c
+    UrlUtils.c
+    CommonUtils.cpp)
+
 target_compile_options(commonutils PRIVATE -Wno-psabi)
+
 target_include_directories(commonutils
     PUBLIC
         ${CMAKE_CURRENT_SOURCE_DIR}
         ${RAPIDJSON_INCLUDE_DIRS}
 )
+
 target_link_libraries(commonutils PRIVATE logging)
