+++ src/common/commonutils/Internal.h
@@ -1,0 +1,44 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#ifndef INTERNAL_H
+#define INTERNAL_H
+
+#define _GNU_SOURCE
+#include <stdio.h>
+#include <stdlib.h>
+#include <stdbool.h>
+#include <stdarg.h>
+#include <string.h>
+#include <errno.h>
+#include <sys/types.h>
+#include <sys/wait.h>
+#include <sys/stat.h>
+#include <sys/file.h>
+#include <unistd.h>
+#include <ctype.h>
+#include <time.h>
+#include <Logging.h>
+#include <CommonUtils.h>
+
+#if ((__GLIBC__ == 2) && (__GLIBC_MINOR__ < 30))
+#include <sys/syscall.h>
+#define gettid() syscall(SYS_gettid)
+#endif
+
+#ifdef __cplusplus
+
+#include <algorithm>
+#include <cstdio>
+#include <cstring>
+#include <string>
+#include <regex>
+#include <Logging.h>
+#include <CommonUtils.h>
+#include <rapidjson/document.h>
+#include <rapidjson/schema.h>
+#include <rapidjson/stringbuffer.h>
+
+#endif //__cplusplus
+
+#endif // INTERNAL_H
