+++ src/common/commonutils/OtherUtils.c
@@ -1,0 +1,41 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#include "internal.h"
+
+char* DuplicateString(const char* source)
+{
+    int length = 0;
+    char* duplicate = NULL;
+
+    if (NULL == source)
+    {
+        return duplicate;
+    }
+
+    length = (int)strlen(source);
+    duplicate = (char*)malloc(length + 1);
+    if (NULL != duplicate)
+    {
+        memcpy(duplicate, source, length);
+        duplicate[length] = 0;
+    }
+
+    return duplicate;
+}
+
+int SleepMilliseconds(long milliseconds)
+{
+    struct timespec remaining = {0};
+    struct timespec interval = {0}; 
+    
+    if ((milliseconds < 0) || (milliseconds > 999999999))
+    {
+        return EINVAL;
+    }
+    
+    interval.tv_sec = (int)(milliseconds / 1000); 
+    interval.tv_nsec = (milliseconds % 1000) * 1000000;
+
+    return nanosleep(&interval, &remaining);
+}
